requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-environ==0.12.0
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-standardized-errors==0.14.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.1.0
PyYAML==6.0.2
referencing==0.36.1
requests==2.32.3
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
